Reduce header leakage
