+++ blast/runblastsearchworker.h
@@ -24,7 +24,6 @@
 #include <QObject>
 #include <QProcess>
 #include <QString>
-#include "program/globals.h"
 
 //This class carries out the task of running blastn and/or
 //tblastn.
